+++ src/app/(3col)/explore/components/LatestTab/actions.ts
@@ -7,12 +7,12 @@ export const fetchPost = async (path: string) => {
   const queryString = path.split("?")[1] || "";
   const params = new URLSearchParams(queryString);
   const cursor = params.get("cursor") || "";
-  const search = params.get("q") || "";
+  const q = params.get("q") || "";
   const limit = params.get("limit") as string;
 
   const { data: response } = await get<SuccessResponse<PostsResponse>>({
-    url: "/api/posts",
-    queryParams: { cursor, search, limit },
+    url: "/api/posts/search",
+    queryParams: { q, cursor, limit },
     cache: "no-store",
   });
